v2_smart_cloud/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
insightface>=0.7.3
onnxruntime>=1.16.0
customtkinter>=5.2.0
gspread>=5.10.0
oauth2client>=4.1.3
streamlit>=1.31.0
